% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_atoms}
\alias{read_atoms}
\title{Read Atom Geometry}
\usage{
read_atoms(result)
}
\arguments{
\item{result}{A named list where result$inp is the connection to the
file to read from, and result$out is an (initially) empty list to
write parsed data to.}
}
\value{
A named list the same as its argument, with newly parsed
data.
}
\description{
Reads the atom data, i.e. atomic number, charge and position for every
atom listed
}
