% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_comment}
\alias{read_comment}
\title{Read Cubefile header comment}
\usage{
read_comment(result)
}
\arguments{
\item{result}{A named list where result$inp is the connection to the
file to read from, and result$out is an (initially) empty list to
write parsed data to.}
}
\value{
A named list the same as its argument, with newly parsed
data.
}
\description{
Read the first two lines from a cubefile, which are the comment.
This should be called before any other parsing function.
}
