% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_n_atom_and_origin}
\alias{read_n_atom_and_origin}
\title{Read line 3 of A Cubefile}
\usage{
read_n_atom_and_origin(result)
}
\arguments{
\item{result}{A named list where result$inp is the connection to the
file to read from, and result$out is an (initially) empty list to
write parsed data to.}
}
\value{
A named list the same as its argument, with newly parsed
data.
}
\description{
Reads the third line of a cubefile, which contains the number of
atoms and the position of the origin
}
