% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_voxels}
\alias{read_voxels}
\title{Read Voxels}
\usage{
read_voxels(result)
}
\arguments{
\item{result}{A named list where result$inp is the connection to the
file to read from, and result$out is an (initially) empty list to
write parsed data to.}
}
\value{
A named list the same as its argument, with newly parsed
data.
}
\description{
Reads the axis directions and the number of voxels along each one
}
